+++ lib/dart_frog_shared.dart
@@ -9,3 +9,7 @@ export 'exceptions/exceptions.dart';
 export 'exceptions/json_exportable.dart';
 export 'logging/log_handler.dart';
 export 'logging/request_context_details.dart';
+export 'middleware/cors_config.dart';
+export 'middleware/cors_middleware.dart';
+export 'middleware/rate_limit_config.dart';
+export 'middleware/rate_limit_middleware.dart';